mailgun test react/test/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.5
Flask-SQLAlchemy==3.0.4
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
newsapi-python==0.2.7
pymongo==4.4.0
requests==2.31.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
